Component scripts moved to scripts/components folder
