+++ src/scripts/components/tabs.js
